+++ packages/core/src/model/chain/index.ts
@@ -1,8 +1,10 @@
 export * from './local'
 // organize imports in alphabet order to sparse the conflict when adding a new chain
 export * from './arbitrum'
+export * from './astar'
 export * from './avalanche'
 export * from './bsc'
+export * from './boba'
 export * from './cronos'
 export * from './fantom'
 export * from './harmony'