+++ packages/example/playwright/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './constants'
-export * from './log'
-export * from './xpath'