+++ packages/core/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @usedapp/core
 
+## 1.0.2
+
+### Patch Changes
+
+- 37c3b9d: 🪱 Attempt adding network if it isn't added to wallet
+- 083e9b1: Export useBlockNumbers hook as part of internal API
+
+## 1.0.2
+
+### Patch Changes
+
+- 37c3b9d: 🪱 Attempt adding network if it isn't added to wallet
+
+## 1.0.1
+
+### Patch Changes
+
+- 5a16232: Add astar network
+- e978fd9: 🥌 Add static calls
+- 3a108ad: 💤 Add zkSync to known chains
+- 5067120: 🐯 Solve getting lowercase account number from metamask after changing account
+- a40a680: Add Boba network
+
 ## 1.0.0
 
 ### Major Changes